outbox status event and segregated unit tests
