+++ packages/amplify_datastore/example/ios/amplify_datastore_unit_tests/DataStorePluginUnitTests.swift
@@ -17,7 +17,6 @@ import XCTest
 import Amplify
 import Combine
 @testable import AmplifyPlugins
-@testable import AWSPluginsCore
 @testable import amplify_datastore
 
 let testSchema: ModelSchema = ModelSchema.init(name: "Post")
@@ -516,261 +515,5 @@ class DataStorePluginUnitTests: XCTestCase {
                     XCTFail()
                 }
             })
-    }
-    
-    // TODO: Move Hub Tests to New File
-    
-    func test_hub_modelSynced_success() throws {
-        let expect = expectation(description: "listener was invoked")
-        class MockDataStoreHubHandler: DataStoreHubEventStreamHandler {
-            var innerExpect: XCTestExpectation?
-            func setExpectation(outerExpect: XCTestExpectation) {
-                innerExpect = outerExpect
-            }
-            override func sendEvent(flutterEvent: [String : Any]) {
-                XCTAssertEqual(flutterEvent["eventName"] as! String, "DataStore.modelSynced")
-                XCTAssertEqual(flutterEvent["modelName"] as! String, "Blog")
-                XCTAssertEqual(flutterEvent["added"] as! Int, 1)
-                XCTAssertEqual(flutterEvent["deleted"] as! Int, 0)
-                XCTAssertEqual(flutterEvent["updated"] as! Int, 0)
-                XCTAssertEqual(flutterEvent["isDeltaSync"] as! Bool, false)
-                XCTAssertEqual(flutterEvent["isFullSync"] as! Bool, true)
-                innerExpect?.fulfill()
-            }
-            override func sendError(description: String, eventName: String) {
-                XCTFail()
-                innerExpect?.fulfill()
-            }
-        }
-
-        var modelSyncedEventBuilder: ModelSyncedEvent.Builder
-        modelSyncedEventBuilder = ModelSyncedEvent.Builder()
-        modelSyncedEventBuilder.modelName = "Blog"
-        modelSyncedEventBuilder.added = 1
-        modelSyncedEventBuilder.deleted = 0;
-        modelSyncedEventBuilder.updated = 0;
-        modelSyncedEventBuilder.isDeltaSync = false;
-        modelSyncedEventBuilder.isFullSync = true;
-
-        let hubHandler = MockDataStoreHubHandler()
-        // passing the expectation we are awaiting into the enclosed class
-        hubHandler.setExpectation(outerExpect: expect)
-        hubHandler.registerModelsForHub(flutterModels: flutterModelSchemaRegistration)
-        hubHandler.setToken()
-
-        let modelSyncedEventPayload = HubPayload(eventName: HubPayload.EventName.DataStore.modelSynced,
-                                                 data: modelSyncedEventBuilder.build())
-        Amplify.Hub.dispatch(to: .dataStore, payload: modelSyncedEventPayload)
-        // expectation needed to ensure that assertions run on main thread
-        waitForExpectations(timeout: 1.0)
-        // cancellation needed to make sure that Hub token is invalidated to
-        // prevent collisions between tests
-        hubHandler.onCancel(withArguments: nil)
-    }
-    
-    func test_hub_ready_success() throws {
-        let expect = expectation(description: "listener was invoked")
-        class MockDataStoreHubHandler: DataStoreHubEventStreamHandler {
-            var innerExpect: XCTestExpectation?
-            func setExpectation(outerExpect: XCTestExpectation) {
-                innerExpect = outerExpect
-            }
-            override func sendEvent(flutterEvent: [String : Any]) {
-                XCTAssertEqual(flutterEvent["eventName"] as! String, "DataStore.ready")
-                innerExpect?.fulfill()
-            }
-            override func sendError(description: String, eventName: String) {
-                XCTFail()
-                innerExpect?.fulfill()
-            }
-        }
-
-        let hubHandler = MockDataStoreHubHandler()
-        hubHandler.setExpectation(outerExpect: expect)
-        hubHandler.registerModelsForHub(flutterModels: flutterModelSchemaRegistration)
-        hubHandler.setToken()
-       
-        let readyEventPayload = HubPayload(eventName: HubPayload.EventName.DataStore.ready)
-        Amplify.Hub.dispatch(to: .dataStore, payload: readyEventPayload)
-        waitForExpectations(timeout: 1.0)
-        hubHandler.onCancel(withArguments: nil)
-
-    }
-    
-    func test_hub_subscriptionEstablished_success() throws {
-        let expect = expectation(description: "listener was invoked")
-        class MockDataStoreHubHandler: DataStoreHubEventStreamHandler {
-            var innerExpect: XCTestExpectation?
-            func setExpectation(outerExpect: XCTestExpectation) {
-                innerExpect = outerExpect
-            }
-            override func sendEvent(flutterEvent: [String : Any]) {
-                XCTAssertEqual(flutterEvent["eventName"] as! String, "DataStore.subscriptionEstablished")
-                innerExpect?.fulfill()
-            }
-            override func sendError(description: String, eventName: String) {
-                XCTFail()
-                innerExpect?.fulfill()
-            }
-        }
-
-        let hubHandler = MockDataStoreHubHandler()
-        hubHandler.setExpectation(outerExpect: expect)
-        hubHandler.registerModelsForHub(flutterModels: flutterModelSchemaRegistration)
-        hubHandler.setToken()
-
-        let subscriptionEstablishedPayload = HubPayload(eventName: HubPayload.EventName.DataStore.subscriptionsEstablished)
-        Amplify.Hub.dispatch(to: .dataStore, payload: subscriptionEstablishedPayload)
-        waitForExpectations(timeout: 1.0)
-        hubHandler.onCancel(withArguments: nil)
-    }
-
-    func test_hub_syncQueriesReady_success() throws {
-        let expect = expectation(description: "listener was invoked")
-        class MockDataStoreHubHandler: DataStoreHubEventStreamHandler {
-            var innerExpect: XCTestExpectation?
-            func setExpectation(outerExpect: XCTestExpectation) {
-                innerExpect = outerExpect
-            }
-            override func sendEvent(flutterEvent: [String : Any]) {
-                XCTAssertEqual(flutterEvent["eventName"] as! String, "DataStore.syncQueriesReady")
-                innerExpect?.fulfill()
-            }
-            override func sendError(description: String, eventName: String) {
-                XCTFail()
-                innerExpect?.fulfill()
-            }
-        }
-
-        let hubHandler = MockDataStoreHubHandler()
-        hubHandler.setExpectation(outerExpect: expect)
-        hubHandler.registerModelsForHub(flutterModels: flutterModelSchemaRegistration)
-        hubHandler.setToken()
-       
-        let syncQueriesReadyPayload = HubPayload(eventName: HubPayload.EventName.DataStore.syncQueriesReady)
-        Amplify.Hub.dispatch(to: .dataStore, payload: syncQueriesReadyPayload)
-        waitForExpectations(timeout: 1.0)
-        hubHandler.onCancel(withArguments: nil)
-    }
-    
-    func test_hub_networkStatusEvent_success() throws {
-        let expect = expectation(description: "listener was invoked")
-        class MockDataStoreHubHandler: DataStoreHubEventStreamHandler {
-            var innerExpect: XCTestExpectation?
-            func setExpectation(outerExpect: XCTestExpectation) {
-                innerExpect = outerExpect
-            }
-            override func sendEvent(flutterEvent: [String : Any]) {
-                XCTAssertEqual(flutterEvent["eventName"] as! String, "DataStore.networkStatus")
-                XCTAssertEqual(flutterEvent["active"] as? Bool, true)
-                innerExpect?.fulfill()
-            }
-            override func sendError(description: String, eventName: String) {
-                XCTFail()
-                innerExpect?.fulfill()
-            }
-        }
-        
-        let networkStatusEvent = NetworkStatusEvent(active: true)
-
-        let hubHandler = MockDataStoreHubHandler()
-        hubHandler.setExpectation(outerExpect: expect)
-        hubHandler.registerModelsForHub(flutterModels: flutterModelSchemaRegistration)
-        hubHandler.setToken()
-       
-        let networkStatusPayload = HubPayload(eventName: HubPayload.EventName.DataStore.networkStatus, data: networkStatusEvent)
-        Amplify.Hub.dispatch(to: .dataStore, payload: networkStatusPayload)
-        waitForExpectations(timeout: 1.0)
-        hubHandler.onCancel(withArguments: nil)
-    }
-    
-    func test_hub_syncQueriesStartedEvent_success() throws {
-        let expect = expectation(description: "listener was invoked")
-        class MockDataStoreHubHandler: DataStoreHubEventStreamHandler {
-            var innerExpect: XCTestExpectation?
-            func setExpectation(outerExpect: XCTestExpectation) {
-                innerExpect = outerExpect
-            }
-            override func sendEvent(flutterEvent: [String : Any]) {
-                XCTAssertEqual(flutterEvent["eventName"] as! String, "DataStore.syncQueriesStarted")
-                XCTAssertEqual(flutterEvent["models"] as! String, "[\"Blog\"]")
-                innerExpect?.fulfill()
-            }
-            override func sendError(description: String, eventName: String) {
-                XCTFail()
-                innerExpect?.fulfill()
-            }
-        }
-        
-        let syncQueriesStartedEvent = SyncQueriesStartedEvent(models: ["Blog"])
-        let hubHandler = MockDataStoreHubHandler()
-        hubHandler.setExpectation(outerExpect: expect)
-        hubHandler.registerModelsForHub(flutterModels: flutterModelSchemaRegistration)
-        hubHandler.setToken()
-       
-        let syncQueriesStartedPayload = HubPayload(eventName: HubPayload.EventName.DataStore.syncQueriesStarted, data: syncQueriesStartedEvent)
-        Amplify.Hub.dispatch(to: .dataStore, payload: syncQueriesStartedPayload)
-        waitForExpectations(timeout: 1.0)
-        hubHandler.onCancel(withArguments: nil)
-    }
-    
-    func test_hub_outboxMutationProcessed_success() throws {
-        let expect = expectation(description: "listener was invoked")
-        class MockDataStoreHubHandler: DataStoreHubEventStreamHandler {
-            var innerExpect: XCTestExpectation?
-            func setExpectation(outerExpect: XCTestExpectation) {
-                innerExpect = outerExpect
-            }
-            override func sendEvent(flutterEvent: [String : Any]) {
-                let element = flutterEvent["element"] as! [String : Any]
-                let model = element["model"] as! [String : Any]
-                let serializedData = model["serializedData"] as! [String : Any]
-                XCTAssertEqual(flutterEvent["eventName"] as! String, "DataStore.outboxMutationProcessed")
-                XCTAssertEqual(flutterEvent["modelName"] as! String, "Post")
-                XCTAssertEqual(element["_lastChangedAt"] as? Int, 123)
-                XCTAssertEqual(element["_version"] as? Int, 1)
-                XCTAssertEqual(element["_deleted"] as? Bool, false)
-                XCTAssertEqual(model["modelName"] as! String, "Post")
-                XCTAssertEqual(model["modelName"] as! String, "Post")
-                XCTAssertEqual(serializedData["title"] as! String, "Title 1")
-                innerExpect?.fulfill()
-            }
-            override func sendError(description: String, eventName: String) {
-                XCTFail()
-                innerExpect?.fulfill()
-            }
-        }
-        
-        let uuid =  UUID().uuidString
-        let modelMap = [
-            "modelName": "Post",
-            "id": uuid,
-            "title": "Title 1"
-        ]
-        
-        let serializedModel = FlutterSerializedModel(id: uuid, map: try FlutterDataStoreRequestUtils.getJSONValue(modelMap))
-        
-        let syncMetadata = MutationSyncMetadata(id: uuid,
-                                                deleted: false,
-                                                lastChangedAt: 123,
-                                                version: 1)
-        
-        let hubHandler = MockDataStoreHubHandler()
-        
-        do {
-            var anyModel = try serializedModel.eraseToAnyModel()
-            let mutationSync = MutationSync(model: anyModel, syncMetadata: syncMetadata)
-            let outboxMutationProcessedEvent = OutboxMutationEvent.fromModelWithMetadata(modelName: "Post", model: serializedModel, mutationSync: mutationSync)
-            hubHandler.setExpectation(outerExpect: expect)
-            hubHandler.registerModelsForHub(flutterModels: flutterModelSchemaRegistration)
-            hubHandler.setToken()
-            let outboxMutationProcessedPayload = HubPayload(eventName: HubPayload.EventName.DataStore.outboxMutationProcessed, data: outboxMutationProcessedEvent)
-            Amplify.Hub.dispatch(to: .dataStore, payload: outboxMutationProcessedPayload)
-        } catch {
-            XCTFail()
-            expect.fulfill()
-        }
-        waitForExpectations(timeout: 1.0)
-        hubHandler.onCancel(withArguments: nil)
     }
 }
